cleanup/reorg of utils
