+++ sdk/utils.go
@@ -4,12 +4,10 @@ import (
 	"crypto/md5" // #nosec
 	"fmt"
 	"io"
-	"os"
 	"strings"
 	"time"
 
 	"github.com/vapor-ware/synse-sdk/sdk/config"
-	"github.com/vapor-ware/synse-sdk/sdk/logger"
 )
 
 // makeIDString makes a compound string out of the given rack, board, and
@@ -17,134 +15,6 @@ import (
 // for a given device.
 func makeIDString(rack, board, device string) string {
 	return strings.Join([]string{rack, board, device}, "-")
-}
-
-// getHandlerForDevice gets the DeviceHandler for a device, based on the handler name.
-func getHandlerForDevice(handlerName string) (*DeviceHandler, error) {
-	for _, handler := range deviceHandlers {
-		if handler.Name == handlerName {
-			return handler, nil
-		}
-	}
-	return nil, fmt.Errorf("no handler found with name: %s", handlerName)
-}
-
-// makeDevices creates Device instances from a DeviceConfig. The DeviceConfig
-// used here should be a unified config, meaning that all DeviceConfigs (either from
-// different files or from file and dynamic registration) are merged into a single
-// DeviceConfig. This should only be called once all configs have been parsed and
-// validated to ensure that the information we have is all correct.
-func makeDevices(config *config.DeviceConfig) ([]*Device, error) {
-	var devices []*Device
-
-	// The DeviceConfig we get here should be the unified config.
-	for _, kind := range config.Devices {
-		for _, instance := range kind.Instances {
-
-			// Get the outputs for the instance.
-			instanceOutputs, err := getInstanceOutputs(kind, instance)
-
-			// Get the location
-			l, err := config.GetLocation(instance.Location)
-			if err != nil {
-				return nil, err
-			}
-			location, err := NewLocationFromConfig(l)
-			if err != nil {
-				return nil, err
-			}
-
-			// Get the DeviceHandler. If a specific handlerName is set in the config,
-			// we will use that as the definitive handler. Otherwise, use the kind.
-			handlerName := kind.Name
-			if kind.HandlerName != "" {
-				handlerName = kind.HandlerName
-			}
-			if instance.HandlerName != "" {
-				handlerName = instance.HandlerName
-			}
-			handler, err := getHandlerForDevice(handlerName)
-			if err != nil {
-				return nil, err
-			}
-
-			device := &Device{
-				// The name of the device kind. This is essentially the identifier
-				// for the device type.
-				Kind: kind.Name,
-
-				// Any metadata associated with the device kind.
-				Metadata: kind.Metadata,
-
-				// The name of the plugin.
-				Plugin: metainfo.Name,
-
-				// Device-level information. This is not output-specific info.
-				Info: instance.Info,
-
-				// The location of the device.
-				Location: location,
-
-				// Any data associated with the device instance.
-				Data: instance.Data,
-
-				// The outputs supported by the device. A device output may
-				// supply more info, like Data, Info, Type, etc, so that should
-				// be accounted for when doing readings/writing stuff..
-				Outputs: instanceOutputs,
-
-				// The read/write handler for the device. Handlers should be registered globally.
-				Handler: handler,
-			}
-
-			devices = append(devices, device)
-		}
-	}
-	return devices, nil
-}
-
-// getInstanceOutputs get the Outputs for a single device instance. It converts
-// the instance's DeviceOutput to an Output type, and by doing so unifies that
-// output with its corresponding OutputType information.
-//
-// If output inheritance is enable for the instance (which is it by default),
-// this will also take the DeviceOutputs defined by the instance's kind.
-func getInstanceOutputs(kind *config.DeviceKind, instance *config.DeviceInstance) ([]*Output, error) {
-	var instanceOutputs []*Output
-
-	// Create the outputs specific to the instance first.
-	for _, o := range instance.Outputs {
-		output, err := NewOutputFromConfig(o)
-		if err != nil {
-			return nil, err
-		}
-		instanceOutputs = append(instanceOutputs, output)
-	}
-
-	// If output inheritance is not disabled, we will take any outputs
-	// from the DeviceKind as well. If there is an output with the same
-	// name already set from the instance config, we will ignore it.
-	if !instance.DisableOutputInheritance {
-		for _, o := range kind.Outputs {
-			output, err := NewOutputFromConfig(o)
-			if err != nil {
-				return nil, err
-			}
-			// Check if the output is already being tracked
-			duplicate := false
-			for _, tracked := range instanceOutputs {
-				if tracked.Name == output.Name {
-					duplicate = true
-					break
-				}
-			}
-			if !duplicate {
-				instanceOutputs = append(instanceOutputs, output)
-			}
-		}
-	}
-
-	return instanceOutputs, nil
 }
 
 // getTypeByName gets the output type with the given name. If an output type does
@@ -155,34 +25,6 @@ func getTypeByName(name string) (*config.OutputType, error) {
 		return nil, fmt.Errorf("no output type with name '%s' found", name)
 	}
 	return t, nil
-}
-
-// setupSocket is used to make sure the path for unix socket used for gRPC communication
-// is set up and accessible locally. Creates the directory for the socket. Returns the
-// directoryName and err.
-func setupSocket(name string) (string, error) {
-	socket := fmt.Sprintf("%s/%s", sockPath, name)
-
-	_, err := os.Stat(sockPath)
-	if err != nil {
-		if os.IsNotExist(err) {
-			if err = os.MkdirAll(sockPath, os.ModePerm); err != nil {
-				return "", err
-			}
-		} else {
-			logger.Errorf("failed to create socket path %v: %v", sockPath, err)
-			return "", err
-		}
-	} else {
-		err = os.Remove(socket)
-		if err != nil {
-			if !os.IsNotExist(err) {
-				logger.Errorf("failed to remove existing socket %v: %v", socket, err)
-				return "", err
-			}
-		}
-	}
-	return socket, nil
 }
 
 // newUID creates a new unique identifier for a Device. This id should be
